# public/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Use API to call the music generation service of Suno.ai and easily integrate it into agents like GPTs.\n\n**Multi-account**: every endpoint accepts `account` (body/form/query) or the `x-suno-account` header (account id, label, email or handle) to pick the Suno account. Without it the default account is used. `suno_cookie` still overrides everything.",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/api/uploads/audio/{audio_id}/initialize-clip": {
      "post": {
        "summary": "Initialize an uploaded audio as a clip.",
        "description": "Wraps Suno's upload initialization endpoint and returns the generated `clip_id` for a completed uploaded audio.",
        "tags": [
          "default"
        ],
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-suno-account",
            "in": "header",
            "required": false,
            "description": "Suno account to use (id, label, email or handle). Default account when omitted.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clip_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/playlist/create": {
      "post": {
        "summary": "Create a playlist.",
        "description": "Creates a playlist in one exposed API call. Internally it performs Suno's draft creation and metadata update so the caller only needs to provide the final name.",
        "tags": [
          "default"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "parameters": [
          {
            "name": "x-suno-account",
            "in": "header",
            "required": false,
            "description": "Suno account to use (id, label, email or handle). Default account when omitted.",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/workspace_feed": {
      "get": {
        "summary": "Get the feed for a workspace.",
        "description": "Returns feed/clip information for the requested workspace. The workspace can be identified by id or by exact name.",
        "tags": [
          "default"
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "x-suno-account",
            "in": "header",
            "required": false,
            "description": "Suno account to use (id, label, email or handle). Default account when omitted.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workspace": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    },
                    "clips": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "suno model name, chirp-v3"
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Required when SUNO_API_KEY is set on the server"
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}